+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+### v4.2.0
+*  FIX (changelog): update changelog [View](https://github.com/sanusart/react-dropdown-select/commit/cee760c1376ba684456765a8ff6dd2ef3e45e4a5)
+*  Fixed an issue where top positioning would not work in portal mode (#77) closes #73 Thanks to @akdjr [View](https://github.com/sanusart/react-dropdown-select/commit/d7dc82d9261f9a4ed636c5b967b99fe72d1210f5)
+*  v4.1.0 See changelog: https://github.com/sanusart/react-dropdown-select/blob/master/CHANGELOG.md [View](https://github.com/sanusart/react-dropdown-select/commit/e0ee950cd80da69adfed30ee27b435126758da21)
+*  FIX (types): types improvement by @435352980 closes #76 [View](https://github.com/sanusart/react-dropdown-select/commit/8c9f47dec1201c9c5de0eee0e34968442b037a3e)
+*  v4.1.1 See changelog: https://github.com/sanusart/react-dropdown-select/blob/master/CHANGELOG.md [View](https://github.com/sanusart/react-dropdown-select/commit/220dfcbd99f07a8ee62971dbd356786b84da3ac8)
+*  FEATURE (props): add onDropdownCloseRequest prop to notify about clos… (#80) [View](https://github.com/sanusart/react-dropdown-select/commit/c5363474f002852a07ae713e8fb335270d2dfc78)
+
+### v4.1.1
+*  FIX (types): types improvement by @435352980 closes #76 [View](https://github.com/sanusart/react-dropdown-select/commit/8c9f47dec1201c9c5de0eee0e34968442b037a3e)
+
+### v4.1.0
+*  Fixed an issue where top positioning would not work in portal mode (#77) closes #73 Thanks to @akdjr [View](https://github.com/sanusart/react-dropdown-select/commit/d7dc82d9261f9a4ed636c5b967b99fe72d1210f5)
+
+### v4.0.1
+*  FIX (props): fix prop type warnings (#75) thanks to @kkkrist [View](https://github.com/sanusart/react-dropdown-select/commit/b7429cfe1a960ae84eea4c25663803545a8c6b8a)
+
+### v4.0.0
+*  FIX (ssr) [#66]: incompatible with SSR closes #66 [View](https://github.com/sanusart/react-dropdown-select/commit/e1a882dae3a298a75ad9cc6b6c11810a48a8cae8)
+
+### v3.12.0
+*  FEATURE (events) [#70]: Remove selected option on key press of backspace button, closes #70 (#71) [View](https://github.com/sanusart/react-dropdown-select/commit/30a2aee405af2981eaa3aeb2a246f1d46f3647af)
+
 ### v3.11.1
 *  FIX (types)[#68]: wrong import and portal type in types.d.ts closes #68 (#69) [View](https://github.com/sanusart/react-dropdown-select/commit/3dfdb2d21e3d446c73e36f97d5d02f558be3976d)